Fix bug switching to sf3.*

+++ Entity/Traits/BaseAdminEntityRepository.php
@@ -5,15 +5,17 @@ namespace Cms\Bundle\AdminBundle\Entity\Traits;
 use Doctrine\ORM\Query;
 
 trait BaseAdminEntityRepository {
-	
-	public function deleteGroup($ids) {
-        $qb = $this->createQueryBuilder('t');
 
-        $qb->delete()
-			->where($qb->expr()->in('t.id', $ids))
-			->getQuery()
-			->execute(null, Query::HYDRATE_OBJECT);
+    public function deleteGroup($ids) {
+        if ($ids) {
+            $qb = $this->createQueryBuilder('t');
 
-        $this->_em->flush();
+            $qb->delete()
+                ->where($qb->expr()->in('t.id', $ids))
+                ->getQuery()
+                ->execute(null, Query::HYDRATE_OBJECT);
+
+            $this->getEntityManager()->flush();
+        }
     }
 }
